feat: use storable messages instead of forced ping-pong
Previously, the server sent a message to the client, and could only send another when the client sent a message, this was causing problems in scenarios with a lot of packet loss (above 20%), because there simply wasn't enough time for the packets to be sent and confirmed. With this approach, the server continues sending messages to the client, even if the client does not respond, and the client continues sending messages to the server. The number of messages sent by the server and the client remains the same (1:1), but now they do not necessarily need to be sent alternately.