feat: enhanced Client::disconnect return
More information of what happened at the client itself disconnection.
Why the runtime is returned, but with private field? Well, basically, in the actual runtime (tokio) can not be dropped inside it own async task, returning the runtime (inside Arc<>) asserts that the runtime will live at least at the end of the client disconnection, without the lifetimes.