feat(examples/bevy-pong): rename from bevy-ping-pong
To bevy-pong.
The real name of this game is not ping-pong, is pong (https://pt.wikipedia.org/wiki/Pong).
No additional reasons, just confusion.
A rebase would be possible to rename the commits, but since this example has been there since the beginning of the crate's development, the rebase would affect a long time in the past just to rename the commit prefix from "(bevy-ping-pong)" to "(bevy-pong)".