feat: remove reason from ignored ips
Encryption must also be applied here, that is, the entire key exchange layer must be executed. The reason the server could store the reason the IP was ignored was to gain performance, but since encryption must be applied, the performance gain is small, and the same functionality can be achieved more simply by the common way of refusing clients with Server::refuse.