feat: define constant typing for Packet
In sd_bincode, the Packet type used Deserialize and Serialize (from serde) as a mandatory requirement, now it is not mandatory, only to use the derive of Packet (will come in the next commit).
This will allow essential packets to have the same serialization/deserialization, a hybrid approach for packet consumers will be possible, with packets implemented by derive, and packets implemented manually.