ci: add new workflow for cargo tests
It will be used for continuous integration (CI) and to keep the code safe with tests, guided by Test-Driven Development (TDD).
As there is already a lot of code and many features implemented in the code, incremental TDD or Test-After Development (TAD) will be implemented initially, and after that, as the project already has a base, TDD will be applied to maintain code quality.