feat: remove examples from workspace
Making them isolated projects.
Due to incompatible features, such as different rt_, the code in the main crate does not compile in the development environment, because cargo tries to enable all features, for all members of the workspace. At the time, I (Robsutar) was commenting on the examples that I was not using, this approach is a bit peculiar, better to remove it.